Get a data.frame (in long format) of population by age for multiple countries

Get a data.frame (in long format) of population by age for multiple countries

age_df_countries(countries)

Arguments

countries
A character string or a vector of character containing the names of the countries for which to return contact data

Value

A data.frame (in long format) with 3 columns:

country: the country name

age: the age group

pop: the number of people in this age group

References

https://www.census.gov/programs-surveys/international-programs/about/idb.html

Examples

age_df_countries(c("Austria", "Belgium"))

#>     Country   age     pop
#> 97  Austria 70_74  412613
#> 98  Austria 65_69  463005
#> 99  Austria 60_64  577840
#> 100 Austria   5_9  413258
#> 101 Austria 55_59  699785
#> 102 Austria 50_54  716265
#> 103 Austria 45_49  617300
#> 104 Austria 40_44  556170
#> 105 Austria 35_39  600221
#> 106 Austria 30_34  590607
#> 107 Austria 25_29  583279
#> 108 Austria 20_24  490270
#> 109 Austria 15_19  427882
#> 110 Austria 10_14  407418
#> 111 Austria   0_4  420129
#> 112 Austria   75+  876154
#> 177 Belgium 70_74  559926
#> 178 Belgium 65_69  637315
#> 179 Belgium 60_64  736126
#> 180 Belgium   5_9  675719
#> 181 Belgium 55_59  804893
#> 182 Belgium 50_54  791316
#> 183 Belgium 45_49  765230
#> 184 Belgium 40_44  756669
#> 185 Belgium 35_39  774709
#> 186 Belgium 30_34  769770
#> 187 Belgium 25_29  740608
#> 188 Belgium 20_24  669644
#> 189 Belgium 15_19  643220
#> 190 Belgium 10_14  680427
#> 191 Belgium   0_4  659696
#> 192 Belgium   75+ 1038949