Get the list of countries included in the dataset

Get the list of countries included in the dataset

available_contacts()

Value

A character vector with the name of all countries included in the dataset

Note

This package uses the nomenclature from the countrycode package. If your names differ from the names used here, you should use countrycode as well to update them.

References

Kiesha Prem, Alex R. Cook, Mark Jit, Projecting social contact matrices in 152 countries using contact surveys and demographic data, PLoS Comp. Biol. (2017), doi: 10.1371/journal.pcbi.1005697

Examples

available_contacts()

#>   [1] "Albania"              "Algeria"              "Andorra"             
#>   [4] "Antigua & Barbuda"    "Argentina"            "Armenia"             
#>   [7] "Australia"            "Austria"              "Azerbaijan"          
#>  [10] "Bahamas"              "Bahrain"              "Bangladesh"          
#>  [13] "Belarus"              "Belgium"              "Belize"              
#>  [16] "Benin"                "Bhutan"               "Bolivia"             
#>  [19] "Bosnia"               "Botswana"             "Brazil"              
#>  [22] "Brunei"               "Bulgaria"             "Burkina Faso"        
#>  [25] "Cape Verde"           "Cambodia"             "Cameroon"            
#>  [28] "Canada"               "Chile"                "China"               
#>  [31] "Colombia"             "Congo - Brazzaville"  "Costa Rica"          
#>  [34] "Croatia"              "Cyprus"               "Czechia"             
#>  [37] "Denmark"              "Dominican Republic"   "Ecuador"             
#>  [40] "Egypt"                "El Salvador"          "Estonia"             
#>  [43] "Ethiopia"             "Fiji"                 "Finland"             
#>  [46] "France"               "Georgia"              "Germany"             
#>  [49] "Ghana"                "Greece"               "Guatemala"           
#>  [52] "Guinea"               "Guyana"               "Haiti"               
#>  [55] "Honduras"             "Hong Kong"            "Hungary"             
#>  [58] "Iceland"              "India"                "Indonesia"           
#>  [61] "Iran"                 "Iraq"                 "Ireland"             
#>  [64] "Israel"               "Italy"                "Jamaica"             
#>  [67] "Japan"                "Jordan"               "Kazakhstan"          
#>  [70] "Kenya"                "Kiribati"             "Kuwait"              
#>  [73] "Kyrgyzstan"           "Laos"                 "Latvia"              
#>  [76] "Lebanon"              "Lesotho"              "Liberia"             
#>  [79] "Lithuania"            "Luxembourg"           "Malaysia"            
#>  [82] "Maldives"             "Malta"                "Mauritania"          
#>  [85] "Mauritius"            "Mexico"               "Monaco"              
#>  [88] "Mongolia"             "Montenegro"           "Morocco"             
#>  [91] "Mozambique"           "Namibia"              "Nepal"               
#>  [94] "Netherlands"          "New Zealand"          "Nicaragua"           
#>  [97] "Niger"                "Nigeria"              "Oman"                
#> [100] "Pakistan"             "Panama"               "Paraguay"            
#> [103] "Peru"                 "Philippines"          "Poland"              
#> [106] "Portugal"             "Qatar"                "South Korea"         
#> [109] "Romania"              "Russia"               "Rwanda"              
#> [112] "St. Lucia"            "Samoa"                "São Tomé & Príncipe" 
#> [115] "Saudi Arabia"         "Senegal"              "Serbia"              
#> [118] "Seychelles"           "Sierra Leone"         "Singapore"           
#> [121] "Slovakia"             "Slovenia"             "Solomon Islands"     
#> [124] "South Africa"         "Spain"                "Sri Lanka"           
#> [127] "Suriname"             "Sweden"               "Switzerland"         
#> [130] "Syria"                "Taiwan"               "Tajikistan"          
#> [133] "North Macedonia"      "Thailand"             "Timor-Leste"         
#> [136] "Tonga"                "Tunisia"              "Turkey"              
#> [139] "Uganda"               "Ukraine"              "United Arab Emirates"
#> [142] "UK"                   "Tanzania"             "US"                  
#> [145] "Uruguay"              "Uzbekistan"           "Vanuatu"             
#> [148] "Venezuela"            "Vietnam"              "Yemen"               
#> [151] "Zambia"               "Zimbabwe"