Get a data.frame (in long format) of contact data for multiple countries

Get a data.frame (in long format) of contact data for multiple countries

contact_df_countries(countries, ...)

Arguments

countries

A character string or a vector of character containing the names of the countries for which to return contact data

...

Arguments passed to contact_matrix()

Value

A data.frame (in long format) with 4 columns:

country: the country name

age_from: the age group of individual

age_to: the age group of contact

contact: the intensity of contact

References

Kiesha Prem, Alex R. Cook, Mark Jit, Projecting social contact matrices in 152 countries using contact surveys and demographic data, PLoS Comp. Biol. (2017), doi:10.1371/journal.pcbi.1005697

Kiesha Prem, Kevin van Zandvoort, Petra Klepac, Rosalind M. Eggo, Nicholas G. Davies, CMMID COVID-19 Working Group, Alex R. Cook, Mark Jit, Projecting contact matrices in 177 geographical regions: An update and comparison with empirical data for the COVID-19 era, PLoS Comp. Biol. (2021), doi:10.1371/journal.pcbi.1009098 .

Examples

contact_df_countries(c("Austria", "Belgium"), location = "all")
#>     country age_from  age_to  contact
#> 1   Austria  [00,05) [00,05)  2.57832
#> 2   Austria  [05,10) [00,05)  1.21139
#> 3   Austria  [10,15) [00,05)  0.35961
#> 4   Austria  [15,20) [00,05)  0.28192
#> 5   Austria  [20,25) [00,05)  0.33834
#> 6   Austria  [25,30) [00,05)  0.65584
#> 7   Austria  [30,35) [00,05)  0.78240
#> 8   Austria  [35,40) [00,05)  0.72109
#> 9   Austria  [40,45) [00,05)  0.48626
#> 10  Austria  [45,50) [00,05)  0.31847
#> 11  Austria  [50,55) [00,05)  0.27336
#> 12  Austria  [55,60) [00,05)  0.34435
#> 13  Austria  [60,65) [00,05)  0.32839
#> 14  Austria  [65,70) [00,05)  0.26060
#> 15  Austria  [70,75) [00,05)  0.19196
#> 16  Austria  [75,80) [00,05)  0.25425
#> 17  Austria  [00,05) [05,10)  0.92094
#> 18  Austria  [05,10) [05,10)  5.54205
#> 19  Austria  [10,15) [05,10)  1.92530
#> 20  Austria  [15,20) [05,10)  0.44065
#> 21  Austria  [20,25) [05,10)  0.33149
#> 22  Austria  [25,30) [05,10)  0.39638
#> 23  Austria  [30,35) [05,10)  0.75155
#> 24  Austria  [35,40) [05,10)  0.93742
#> 25  Austria  [40,45) [05,10)  0.66768
#> 26  Austria  [45,50) [05,10)  0.36811
#> 27  Austria  [50,55) [05,10)  0.30513
#> 28  Austria  [55,60) [05,10)  0.34792
#> 29  Austria  [60,65) [05,10)  0.29175
#> 30  Austria  [65,70) [05,10)  0.29123
#> 31  Austria  [70,75) [05,10)  0.26336
#> 32  Austria  [75,80) [05,10)  0.24141
#> 33  Austria  [00,05) [10,15)  0.47330
#> 34  Austria  [05,10) [10,15)  1.04480
#> 35  Austria  [10,15) [10,15)  8.65137
#> 36  Austria  [15,20) [10,15)  2.90195
#> 37  Austria  [20,25) [10,15)  0.38014
#> 38  Austria  [25,30) [10,15)  0.31214
#> 39  Austria  [30,35) [10,15)  0.54276
#> 40  Austria  [35,40) [10,15)  0.71856
#> 41  Austria  [40,45) [10,15)  0.91165
#> 42  Austria  [45,50) [10,15)  0.59656
#> 43  Austria  [50,55) [10,15)  0.51753
#> 44  Austria  [55,60) [10,15)  0.44696
#> 45  Austria  [60,65) [10,15)  0.28614
#> 46  Austria  [65,70) [10,15)  0.25578
#> 47  Austria  [70,75) [10,15)  0.29514
#> 48  Austria  [75,80) [10,15)  0.28626
#> 49  Austria  [00,05) [15,20)  0.36082
#> 50  Austria  [05,10) [15,20)  0.40085
#> 51  Austria  [10,15) [15,20)  0.92765
#> 52  Austria  [15,20) [15,20)  9.12983
#> 53  Austria  [20,25) [15,20)  1.86793
#> 54  Austria  [25,30) [15,20)  0.71920
#> 55  Austria  [30,35) [15,20)  0.43897
#> 56  Austria  [35,40) [15,20)  0.68920
#> 57  Austria  [40,45) [15,20)  0.80602
#> 58  Austria  [45,50) [15,20)  1.00036
#> 59  Austria  [50,55) [15,20)  0.77494
#> 60  Austria  [55,60) [15,20)  0.49052
#> 61  Austria  [60,65) [15,20)  0.32093
#> 62  Austria  [65,70) [15,20)  0.24999
#> 63  Austria  [70,75) [15,20)  0.38040
#> 64  Austria  [75,80) [15,20)  0.28610
#> 65  Austria  [00,05) [20,25)  0.49986
#> 66  Austria  [05,10) [20,25)  0.33672
#> 67  Austria  [10,15) [20,25)  0.44340
#> 68  Austria  [15,20) [20,25)  1.16607
#> 69  Austria  [20,25) [20,25)  3.36886
#> 70  Austria  [25,30) [20,25)  1.66501
#> 71  Austria  [30,35) [20,25)  0.92077
#> 72  Austria  [35,40) [20,25)  0.72108
#> 73  Austria  [40,45) [20,25)  0.86495
#> 74  Austria  [45,50) [20,25)  0.90108
#> 75  Austria  [50,55) [20,25)  1.00680
#> 76  Austria  [55,60) [20,25)  0.70345
#> 77  Austria  [60,65) [20,25)  0.47466
#> 78  Austria  [65,70) [20,25)  0.36032
#> 79  Austria  [70,75) [20,25)  0.30593
#> 80  Austria  [75,80) [20,25)  0.26106
#> 81  Austria  [00,05) [25,30)  0.88350
#> 82  Austria  [05,10) [25,30)  0.70215
#> 83  Austria  [10,15) [25,30)  0.51938
#> 84  Austria  [15,20) [25,30)  0.70140
#> 85  Austria  [20,25) [25,30)  1.46198
#> 86  Austria  [25,30) [25,30)  3.02566
#> 87  Austria  [30,35) [25,30)  1.55400
#> 88  Austria  [35,40) [25,30)  1.25972
#> 89  Austria  [40,45) [25,30)  1.11388
#> 90  Austria  [45,50) [25,30)  0.98343
#> 91  Austria  [50,55) [25,30)  1.30587
#> 92  Austria  [55,60) [25,30)  1.13902
#> 93  Austria  [60,65) [25,30)  0.75194
#> 94  Austria  [65,70) [25,30)  0.53820
#> 95  Austria  [70,75) [25,30)  0.44956
#> 96  Austria  [75,80) [25,30)  0.30864
#> 97  Austria  [00,05) [30,35)  1.19903
#> 98  Austria  [05,10) [30,35)  1.03313
#> 99  Austria  [10,15) [30,35)  0.65650
#> 100 Austria  [15,20) [30,35)  0.57586
#> 101 Austria  [20,25) [30,35)  0.91134
#> 102 Austria  [25,30) [30,35)  1.59066
#> 103 Austria  [30,35) [30,35)  2.49274
#> 104 Austria  [35,40) [30,35)  1.55811
#> 105 Austria  [40,45) [30,35)  1.41692
#> 106 Austria  [45,50) [30,35)  1.15213
#> 107 Austria  [50,55) [30,35)  1.12041
#> 108 Austria  [55,60) [30,35)  1.18836
#> 109 Austria  [60,65) [30,35)  0.82714
#> 110 Austria  [65,70) [30,35)  0.71016
#> 111 Austria  [70,75) [30,35)  0.47634
#> 112 Austria  [75,80) [30,35)  0.46043
#> 113 Austria  [00,05) [35,40)  1.08644
#> 114 Austria  [05,10) [35,40)  1.19617
#> 115 Austria  [10,15) [35,40)  1.03175
#> 116 Austria  [15,20) [35,40)  0.85259
#> 117 Austria  [20,25) [35,40)  0.75262
#> 118 Austria  [25,30) [35,40)  1.18200
#> 119 Austria  [30,35) [35,40)  1.58806
#> 120 Austria  [35,40) [35,40)  2.61579
#> 121 Austria  [40,45) [35,40)  1.62128
#> 122 Austria  [45,50) [35,40)  1.31220
#> 123 Austria  [50,55) [35,40)  1.09831
#> 124 Austria  [55,60) [35,40)  0.96128
#> 125 Austria  [60,65) [35,40)  0.86923
#> 126 Austria  [65,70) [35,40)  0.69605
#> 127 Austria  [70,75) [35,40)  0.69387
#> 128 Austria  [75,80) [35,40)  0.46534
#> 129 Austria  [00,05) [40,45)  0.64976
#> 130 Austria  [05,10) [40,45)  1.01740
#> 131 Austria  [10,15) [40,45)  1.21721
#> 132 Austria  [15,20) [40,45)  0.98825
#> 133 Austria  [20,25) [40,45)  0.71323
#> 134 Austria  [25,30) [40,45)  0.96930
#> 135 Austria  [30,35) [40,45)  1.14880
#> 136 Austria  [35,40) [40,45)  1.69460
#> 137 Austria  [40,45) [40,45)  2.36411
#> 138 Austria  [45,50) [40,45)  1.40519
#> 139 Austria  [50,55) [40,45)  1.38226
#> 140 Austria  [55,60) [40,45)  1.01831
#> 141 Austria  [60,65) [40,45)  0.74004
#> 142 Austria  [65,70) [40,45)  0.67864
#> 143 Austria  [70,75) [40,45)  0.75598
#> 144 Austria  [75,80) [40,45)  0.50439
#> 145 Austria  [00,05) [45,50)  0.47854
#> 146 Austria  [05,10) [45,50)  0.58568
#> 147 Austria  [10,15) [45,50)  0.82655
#> 148 Austria  [15,20) [45,50)  1.08290
#> 149 Austria  [20,25) [45,50)  1.01194
#> 150 Austria  [25,30) [45,50)  0.90950
#> 151 Austria  [30,35) [45,50)  0.96049
#> 152 Austria  [35,40) [45,50)  1.17883
#> 153 Austria  [40,45) [45,50)  1.47086
#> 154 Austria  [45,50) [45,50)  2.02146
#> 155 Austria  [50,55) [45,50)  1.60351
#> 156 Austria  [55,60) [45,50)  0.92374
#> 157 Austria  [60,65) [45,50)  0.66507
#> 158 Austria  [65,70) [45,50)  0.51954
#> 159 Austria  [70,75) [45,50)  0.66419
#> 160 Austria  [75,80) [45,50)  0.63777
#> 161 Austria  [00,05) [50,55)  0.54661
#> 162 Austria  [05,10) [50,55)  0.49701
#> 163 Austria  [10,15) [50,55)  0.57475
#> 164 Austria  [15,20) [50,55)  0.70464
#> 165 Austria  [20,25) [50,55)  0.79999
#> 166 Austria  [25,30) [50,55)  1.04626
#> 167 Austria  [30,35) [50,55)  0.90129
#> 168 Austria  [35,40) [50,55)  0.96816
#> 169 Austria  [40,45) [50,55)  1.12059
#> 170 Austria  [45,50) [50,55)  1.25765
#> 171 Austria  [50,55) [50,55)  1.94412
#> 172 Austria  [55,60) [50,55)  1.31226
#> 173 Austria  [60,65) [50,55)  0.73133
#> 174 Austria  [65,70) [50,55)  0.60889
#> 175 Austria  [70,75) [50,55)  0.59034
#> 176 Austria  [75,80) [50,55)  0.71759
#> 177 Austria  [00,05) [55,60)  0.50700
#> 178 Austria  [05,10) [55,60)  0.44970
#> 179 Austria  [10,15) [55,60)  0.39081
#> 180 Austria  [15,20) [55,60)  0.41256
#> 181 Austria  [20,25) [55,60)  0.53849
#> 182 Austria  [25,30) [55,60)  0.68474
#> 183 Austria  [30,35) [55,60)  0.77166
#> 184 Austria  [35,40) [55,60)  0.68418
#> 185 Austria  [40,45) [55,60)  0.57535
#> 186 Austria  [45,50) [55,60)  0.71731
#> 187 Austria  [50,55) [55,60)  1.12574
#> 188 Austria  [55,60) [55,60)  1.72105
#> 189 Austria  [60,65) [55,60)  0.99125
#> 190 Austria  [65,70) [55,60)  0.72362
#> 191 Austria  [70,75) [55,60)  0.56122
#> 192 Austria  [75,80) [55,60)  0.51528
#> 193 Austria  [00,05) [60,65)  0.37891
#> 194 Austria  [05,10) [60,65)  0.39709
#> 195 Austria  [10,15) [60,65)  0.29193
#> 196 Austria  [15,20) [60,65)  0.27856
#> 197 Austria  [20,25) [60,65)  0.30529
#> 198 Austria  [25,30) [60,65)  0.43668
#> 199 Austria  [30,35) [60,65)  0.51623
#> 200 Austria  [35,40) [60,65)  0.56191
#> 201 Austria  [40,45) [60,65)  0.51916
#> 202 Austria  [45,50) [60,65)  0.45797
#> 203 Austria  [50,55) [60,65)  0.60430
#> 204 Austria  [55,60) [60,65)  0.84998
#> 205 Austria  [60,65) [60,65)  1.36237
#> 206 Austria  [65,70) [60,65)  0.78041
#> 207 Austria  [70,75) [60,65)  0.92781
#> 208 Austria  [75,80) [60,65)  0.42903
#> 209 Austria  [00,05) [65,70)  0.29396
#> 210 Austria  [05,10) [65,70)  0.28752
#> 211 Austria  [10,15) [65,70)  0.25984
#> 212 Austria  [15,20) [65,70)  0.21992
#> 213 Austria  [20,25) [65,70)  0.21528
#> 214 Austria  [25,30) [65,70)  0.27370
#> 215 Austria  [30,35) [65,70)  0.31787
#> 216 Austria  [35,40) [65,70)  0.38756
#> 217 Austria  [40,45) [65,70)  0.36825
#> 218 Austria  [45,50) [65,70)  0.32556
#> 219 Austria  [50,55) [65,70)  0.32570
#> 220 Austria  [55,60) [65,70)  0.43104
#> 221 Austria  [60,65) [65,70)  0.61346
#> 222 Austria  [65,70) [65,70)  1.07650
#> 223 Austria  [70,75) [65,70)  0.81967
#> 224 Austria  [75,80) [65,70)  0.52299
#> 225 Austria  [00,05) [70,75)  0.23180
#> 226 Austria  [05,10) [70,75)  0.21371
#> 227 Austria  [10,15) [70,75)  0.23088
#> 228 Austria  [15,20) [70,75)  0.18086
#> 229 Austria  [20,25) [70,75)  0.22974
#> 230 Austria  [25,30) [70,75)  0.21038
#> 231 Austria  [30,35) [70,75)  0.21729
#> 232 Austria  [35,40) [70,75)  0.30339
#> 233 Austria  [40,45) [70,75)  0.30044
#> 234 Austria  [45,50) [70,75)  0.29919
#> 235 Austria  [50,55) [70,75)  0.30021
#> 236 Austria  [55,60) [70,75)  0.28777
#> 237 Austria  [60,65) [70,75)  0.43779
#> 238 Austria  [65,70) [70,75)  0.45180
#> 239 Austria  [70,75) [70,75)  1.15714
#> 240 Austria  [75,80) [70,75)  0.54054
#> 241 Austria  [00,05) [75,80)  0.19012
#> 242 Austria  [05,10) [75,80)  0.18893
#> 243 Austria  [10,15) [75,80)  0.20996
#> 244 Austria  [15,20) [75,80)  0.16357
#> 245 Austria  [20,25) [75,80)  0.20098
#> 246 Austria  [25,30) [75,80)  0.16880
#> 247 Austria  [30,35) [75,80)  0.21169
#> 248 Austria  [35,40) [75,80)  0.21532
#> 249 Austria  [40,45) [75,80)  0.25187
#> 250 Austria  [45,50) [75,80)  0.27342
#> 251 Austria  [50,55) [75,80)  0.28435
#> 252 Austria  [55,60) [75,80)  0.25519
#> 253 Austria  [60,65) [75,80)  0.28647
#> 254 Austria  [65,70) [75,80)  0.30960
#> 255 Austria  [70,75) [75,80)  0.46497
#> 256 Austria  [75,80) [75,80)  0.55061
#> 257 Belgium  [00,05) [00,05)  4.55145
#> 258 Belgium  [05,10) [00,05)  0.99438
#> 259 Belgium  [10,15) [00,05)  0.44314
#> 260 Belgium  [15,20) [00,05)  0.55666
#> 261 Belgium  [20,25) [00,05)  0.20448
#> 262 Belgium  [25,30) [00,05)  0.66245
#> 263 Belgium  [30,35) [00,05)  1.01292
#> 264 Belgium  [35,40) [00,05)  0.79248
#> 265 Belgium  [40,45) [00,05)  0.27695
#> 266 Belgium  [45,50) [00,05)  0.31718
#> 267 Belgium  [50,55) [00,05)  0.34073
#> 268 Belgium  [55,60) [00,05)  0.40043
#> 269 Belgium  [60,65) [00,05)  0.29169
#> 270 Belgium  [65,70) [00,05)  0.15150
#> 271 Belgium  [70,75) [00,05)  0.16239
#> 272 Belgium  [75,80) [00,05)  0.12444
#> 273 Belgium  [00,05) [05,10)  1.77775
#> 274 Belgium  [05,10) [05,10)  7.11763
#> 275 Belgium  [10,15) [05,10)  0.98976
#> 276 Belgium  [15,20) [05,10)  0.92361
#> 277 Belgium  [20,25) [05,10)  0.64295
#> 278 Belgium  [25,30) [05,10)  0.18833
#> 279 Belgium  [30,35) [05,10)  1.07881
#> 280 Belgium  [35,40) [05,10)  0.79677
#> 281 Belgium  [40,45) [05,10)  0.36572
#> 282 Belgium  [45,50) [05,10)  0.19465
#> 283 Belgium  [50,55) [05,10)  0.46882
#> 284 Belgium  [55,60) [05,10)  0.21165
#> 285 Belgium  [60,65) [05,10)  0.14349
#> 286 Belgium  [65,70) [05,10)  0.28870
#> 287 Belgium  [70,75) [05,10)  0.29792
#> 288 Belgium  [75,80) [05,10)  0.12444
#> 289 Belgium  [00,05) [10,15)  0.88551
#> 290 Belgium  [05,10) [10,15)  1.12737
#> 291 Belgium  [10,15) [10,15) 12.16144
#> 292 Belgium  [15,20) [10,15)  1.14699
#> 293 Belgium  [20,25) [10,15)  0.56114
#> 294 Belgium  [25,30) [10,15)  0.15398
#> 295 Belgium  [30,35) [10,15)  0.41235
#> 296 Belgium  [35,40) [10,15)  0.88232
#> 297 Belgium  [40,45) [10,15)  1.14834
#> 298 Belgium  [45,50) [10,15)  0.46945
#> 299 Belgium  [50,55) [10,15)  0.41319
#> 300 Belgium  [55,60) [10,15)  0.21894
#> 301 Belgium  [60,65) [10,15)  0.15713
#> 302 Belgium  [65,70) [10,15)  0.16818
#> 303 Belgium  [70,75) [10,15)  0.40911
#> 304 Belgium  [75,80) [10,15)  0.12444
#> 305 Belgium  [00,05) [15,20)  0.27328
#> 306 Belgium  [05,10) [15,20)  0.38917
#> 307 Belgium  [10,15) [15,20)  1.89044
#> 308 Belgium  [15,20) [15,20) 12.58595
#> 309 Belgium  [20,25) [15,20)  2.13220
#> 310 Belgium  [25,30) [15,20)  0.42219
#> 311 Belgium  [30,35) [15,20)  0.61090
#> 312 Belgium  [35,40) [15,20)  0.43736
#> 313 Belgium  [40,45) [15,20)  0.99557
#> 314 Belgium  [45,50) [15,20)  1.00993
#> 315 Belgium  [50,55) [15,20)  0.50401
#> 316 Belgium  [55,60) [15,20)  0.33212
#> 317 Belgium  [60,65) [15,20)  0.19047
#> 318 Belgium  [65,70) [15,20)  0.15227
#> 319 Belgium  [70,75) [15,20)  0.56974
#> 320 Belgium  [75,80) [15,20)  0.12444
#> 321 Belgium  [00,05) [20,25)  0.42807
#> 322 Belgium  [05,10) [20,25)  0.28223
#> 323 Belgium  [10,15) [20,25)  0.85529
#> 324 Belgium  [15,20) [20,25)  2.10385
#> 325 Belgium  [20,25) [20,25)  5.68727
#> 326 Belgium  [25,30) [20,25)  1.06091
#> 327 Belgium  [30,35) [20,25)  0.89724
#> 328 Belgium  [35,40) [20,25)  0.39765
#> 329 Belgium  [40,45) [20,25)  0.60866
#> 330 Belgium  [45,50) [20,25)  0.94407
#> 331 Belgium  [50,55) [20,25)  1.16257
#> 332 Belgium  [55,60) [20,25)  0.85569
#> 333 Belgium  [60,65) [20,25)  0.36624
#> 334 Belgium  [65,70) [20,25)  0.28577
#> 335 Belgium  [70,75) [20,25)  0.12444
#> 336 Belgium  [75,80) [20,25)  0.12444
#> 337 Belgium  [00,05) [25,30)  1.40238
#> 338 Belgium  [05,10) [25,30)  0.65121
#> 339 Belgium  [10,15) [25,30)  0.43167
#> 340 Belgium  [15,20) [25,30)  0.85326
#> 341 Belgium  [20,25) [25,30)  1.62155
#> 342 Belgium  [25,30) [25,30)  3.49815
#> 343 Belgium  [30,35) [25,30)  1.20592
#> 344 Belgium  [35,40) [25,30)  0.83861
#> 345 Belgium  [40,45) [25,30)  0.93124
#> 346 Belgium  [45,50) [25,30)  1.16782
#> 347 Belgium  [50,55) [25,30)  1.38373
#> 348 Belgium  [55,60) [25,30)  1.21401
#> 349 Belgium  [60,65) [25,30)  0.62040
#> 350 Belgium  [65,70) [25,30)  0.31215
#> 351 Belgium  [70,75) [25,30)  0.36089
#> 352 Belgium  [75,80) [25,30)  0.30976
#> 353 Belgium  [00,05) [30,35)  2.27282
#> 354 Belgium  [05,10) [30,35)  1.47990
#> 355 Belgium  [10,15) [30,35)  0.57425
#> 356 Belgium  [15,20) [30,35)  0.46427
#> 357 Belgium  [20,25) [30,35)  1.39242
#> 358 Belgium  [25,30) [30,35)  1.54203
#> 359 Belgium  [30,35) [30,35)  3.49066
#> 360 Belgium  [35,40) [30,35)  1.37170
#> 361 Belgium  [40,45) [30,35)  0.68859
#> 362 Belgium  [45,50) [30,35)  1.22625
#> 363 Belgium  [50,55) [30,35)  0.80073
#> 364 Belgium  [55,60) [30,35)  1.65412
#> 365 Belgium  [60,65) [30,35)  1.25920
#> 366 Belgium  [65,70) [30,35)  0.36127
#> 367 Belgium  [70,75) [30,35)  0.45786
#> 368 Belgium  [75,80) [30,35)  0.30966
#> 369 Belgium  [00,05) [35,40)  1.10318
#> 370 Belgium  [05,10) [35,40)  1.32127
#> 371 Belgium  [10,15) [35,40)  1.51830
#> 372 Belgium  [15,20) [35,40)  0.99188
#> 373 Belgium  [20,25) [35,40)  0.64767
#> 374 Belgium  [25,30) [35,40)  1.25075
#> 375 Belgium  [30,35) [35,40)  2.13104
#> 376 Belgium  [35,40) [35,40)  2.41292
#> 377 Belgium  [40,45) [35,40)  1.34206
#> 378 Belgium  [45,50) [35,40)  1.68175
#> 379 Belgium  [50,55) [35,40)  0.97062
#> 380 Belgium  [55,60) [35,40)  1.22608
#> 381 Belgium  [60,65) [35,40)  0.68625
#> 382 Belgium  [65,70) [35,40)  0.70549
#> 383 Belgium  [70,75) [35,40)  0.62642
#> 384 Belgium  [75,80) [35,40)  0.12444
#> 385 Belgium  [00,05) [40,45)  0.44643
#> 386 Belgium  [05,10) [40,45)  1.04816
#> 387 Belgium  [10,15) [40,45)  1.45242
#> 388 Belgium  [15,20) [40,45)  1.24341
#> 389 Belgium  [20,25) [40,45)  1.07614
#> 390 Belgium  [25,30) [40,45)  0.76208
#> 391 Belgium  [30,35) [40,45)  1.44945
#> 392 Belgium  [35,40) [40,45)  1.36443
#> 393 Belgium  [40,45) [40,45)  2.70255
#> 394 Belgium  [45,50) [40,45)  1.19563
#> 395 Belgium  [50,55) [40,45)  1.29851
#> 396 Belgium  [55,60) [40,45)  1.35977
#> 397 Belgium  [60,65) [40,45)  1.03869
#> 398 Belgium  [65,70) [40,45)  0.55964
#> 399 Belgium  [70,75) [40,45)  0.86959
#> 400 Belgium  [75,80) [40,45)  0.31498
#> 401 Belgium  [00,05) [45,50)  0.29687
#> 402 Belgium  [05,10) [45,50)  0.63626
#> 403 Belgium  [10,15) [45,50)  0.95621
#> 404 Belgium  [15,20) [45,50)  1.66969
#> 405 Belgium  [20,25) [45,50)  1.30385
#> 406 Belgium  [25,30) [45,50)  1.04627
#> 407 Belgium  [30,35) [45,50)  0.88603
#> 408 Belgium  [35,40) [45,50)  0.80237
#> 409 Belgium  [40,45) [45,50)  1.51396
#> 410 Belgium  [45,50) [45,50)  3.34722
#> 411 Belgium  [50,55) [45,50)  1.36121
#> 412 Belgium  [55,60) [45,50)  1.10269
#> 413 Belgium  [60,65) [45,50)  0.79588
#> 414 Belgium  [65,70) [45,50)  0.17427
#> 415 Belgium  [70,75) [45,50)  1.02930
#> 416 Belgium  [75,80) [45,50)  0.20176
#> 417 Belgium  [00,05) [50,55)  0.58702
#> 418 Belgium  [05,10) [50,55)  0.43770
#> 419 Belgium  [10,15) [50,55)  0.33292
#> 420 Belgium  [15,20) [50,55)  0.65572
#> 421 Belgium  [20,25) [50,55)  0.82869
#> 422 Belgium  [25,30) [50,55)  1.00086
#> 423 Belgium  [30,35) [50,55)  0.58921
#> 424 Belgium  [35,40) [50,55)  0.73229
#> 425 Belgium  [40,45) [50,55)  0.93781
#> 426 Belgium  [45,50) [50,55)  0.95345
#> 427 Belgium  [50,55) [50,55)  2.38719
#> 428 Belgium  [55,60) [50,55)  1.60132
#> 429 Belgium  [60,65) [50,55)  0.57667
#> 430 Belgium  [65,70) [50,55)  0.37238
#> 431 Belgium  [70,75) [50,55)  0.79601
#> 432 Belgium  [75,80) [50,55)  0.45133
#> 433 Belgium  [00,05) [55,60)  0.48969
#> 434 Belgium  [05,10) [55,60)  0.39260
#> 435 Belgium  [10,15) [55,60)  0.38579
#> 436 Belgium  [15,20) [55,60)  0.33477
#> 437 Belgium  [20,25) [55,60)  0.40183
#> 438 Belgium  [25,30) [55,60)  0.54367
#> 439 Belgium  [30,35) [55,60)  0.87530
#> 440 Belgium  [35,40) [55,60)  0.32352
#> 441 Belgium  [40,45) [55,60)  0.29522
#> 442 Belgium  [45,50) [55,60)  0.74380
#> 443 Belgium  [50,55) [55,60)  1.12637
#> 444 Belgium  [55,60) [55,60)  2.52713
#> 445 Belgium  [60,65) [55,60)  0.90068
#> 446 Belgium  [65,70) [55,60)  0.35144
#> 447 Belgium  [70,75) [55,60)  0.35307
#> 448 Belgium  [75,80) [55,60)  0.19170
#> 449 Belgium  [00,05) [60,65)  0.30117
#> 450 Belgium  [05,10) [60,65)  0.35591
#> 451 Belgium  [10,15) [60,65)  0.38491
#> 452 Belgium  [15,20) [60,65)  0.21219
#> 453 Belgium  [20,25) [60,65)  0.23703
#> 454 Belgium  [25,30) [60,65)  0.17246
#> 455 Belgium  [30,35) [60,65)  0.32994
#> 456 Belgium  [35,40) [60,65)  0.36880
#> 457 Belgium  [40,45) [60,65)  0.27021
#> 458 Belgium  [45,50) [60,65)  0.29321
#> 459 Belgium  [50,55) [60,65)  0.88946
#> 460 Belgium  [55,60) [60,65)  0.83942
#> 461 Belgium  [60,65) [60,65)  2.66108
#> 462 Belgium  [65,70) [60,65)  0.59231
#> 463 Belgium  [70,75) [60,65)  1.52835
#> 464 Belgium  [75,80) [60,65)  0.21104
#> 465 Belgium  [00,05) [65,70)  0.26798
#> 466 Belgium  [05,10) [65,70)  0.23359
#> 467 Belgium  [10,15) [65,70)  0.22292
#> 468 Belgium  [15,20) [65,70)  0.21465
#> 469 Belgium  [20,25) [65,70)  0.24456
#> 470 Belgium  [25,30) [65,70)  0.15967
#> 471 Belgium  [30,35) [65,70)  0.36680
#> 472 Belgium  [35,40) [65,70)  0.33389
#> 473 Belgium  [40,45) [65,70)  0.26275
#> 474 Belgium  [45,50) [65,70)  0.46172
#> 475 Belgium  [50,55) [65,70)  0.33746
#> 476 Belgium  [55,60) [65,70)  0.52272
#> 477 Belgium  [60,65) [65,70)  0.88576
#> 478 Belgium  [65,70) [65,70)  1.73583
#> 479 Belgium  [70,75) [65,70)  0.80096
#> 480 Belgium  [75,80) [65,70)  0.28330
#> 481 Belgium  [00,05) [70,75)  0.18374
#> 482 Belgium  [05,10) [70,75)  0.21836
#> 483 Belgium  [10,15) [70,75)  0.17782
#> 484 Belgium  [15,20) [70,75)  0.13613
#> 485 Belgium  [20,25) [70,75)  0.21999
#> 486 Belgium  [25,30) [70,75)  0.13715
#> 487 Belgium  [30,35) [70,75)  0.19817
#> 488 Belgium  [35,40) [70,75)  0.29350
#> 489 Belgium  [40,45) [70,75)  0.27114
#> 490 Belgium  [45,50) [70,75)  0.34058
#> 491 Belgium  [50,55) [70,75)  0.25995
#> 492 Belgium  [55,60) [70,75)  0.45458
#> 493 Belgium  [60,65) [70,75)  0.70353
#> 494 Belgium  [65,70) [70,75)  0.57713
#> 495 Belgium  [70,75) [70,75)  1.98195
#> 496 Belgium  [75,80) [70,75)  0.54199
#> 497 Belgium  [00,05) [75,80)  0.22789
#> 498 Belgium  [05,10) [75,80)  0.15680
#> 499 Belgium  [10,15) [75,80)  0.21199
#> 500 Belgium  [15,20) [75,80)  0.17400
#> 501 Belgium  [20,25) [75,80)  0.21728
#> 502 Belgium  [25,30) [75,80)  0.29170
#> 503 Belgium  [30,35) [75,80)  0.15178
#> 504 Belgium  [35,40) [75,80)  0.17469
#> 505 Belgium  [40,45) [75,80)  0.28539
#> 506 Belgium  [45,50) [75,80)  0.23136
#> 507 Belgium  [50,55) [75,80)  0.38186
#> 508 Belgium  [55,60) [75,80)  0.40637
#> 509 Belgium  [60,65) [75,80)  0.30589
#> 510 Belgium  [65,70) [75,80)  0.35840
#> 511 Belgium  [70,75) [75,80)  0.61615
#> 512 Belgium  [75,80) [75,80)  1.47129