contactdata: Social Contact Matrices for 177 Countries

Data package for the supplementary data in Prem et al. (2017) doi:10.1371/journal.pcbi.1005697 and Prem et al. doi:10.1371/journal.pcbi.1009098 . Provides easy access to contact data for 177 countries, for use in epidemiological, demographic or social sciences research.

See also

Author

Maintainer: Hugo Gruson hugo.gruson+R@normalesup.org (ORCID) [copyright holder]

Other contributors:

Kiesha Prem (ORCID) [data contributor]

Alex Richard Cook (ORCID) [data contributor]

Mark Jit (ORCID) [data contributor]