Get a data.frame (in long format) of population by age for multiple countries

Get a data.frame (in long format) of population by age for multiple countries

age_df_countries(countries)

Arguments

countries

A character string or a vector of character containing the names of the countries for which to return contact data

Value

A data.frame (in long format) with 3 columns:

country: the country name

age: the age group

population: the number of people in this age group

References

https://www.census.gov/programs-surveys/international-programs/about/idb.html

Examples

age_df_countries(c("Austria", "Belgium"))
#>     country   age population
#> 145 Austria 70_74     412613
#> 146 Austria 65_69     463005
#> 147 Austria 60_64     577840
#> 148 Austria   5_9     413258
#> 149 Austria 55_59     699785
#> 150 Austria 50_54     716265
#> 151 Austria 45_49     617300
#> 152 Austria 40_44     556170
#> 153 Austria 35_39     600221
#> 154 Austria 30_34     590607
#> 155 Austria 25_29     583279
#> 156 Austria 20_24     490270
#> 157 Austria 15_19     427882
#> 158 Austria 10_14     407418
#> 159 Austria   0_4     420129
#> 160 Austria   75+     876154
#> 257 Belgium 70_74     559926
#> 258 Belgium 65_69     637315
#> 259 Belgium 60_64     736126
#> 260 Belgium   5_9     675719
#> 261 Belgium 55_59     804893
#> 262 Belgium 50_54     791316
#> 263 Belgium 45_49     765230
#> 264 Belgium 40_44     756669
#> 265 Belgium 35_39     774709
#> 266 Belgium 30_34     769770
#> 267 Belgium 25_29     740608
#> 268 Belgium 20_24     669644
#> 269 Belgium 15_19     643220
#> 270 Belgium 10_14     680427
#> 271 Belgium   0_4     659696
#> 272 Belgium   75+    1038949