Get a data.frame (in long format) of contact data for multiple countries

Get a data.frame (in long format) of contact data for multiple countries

contact_df_countries(
  countries,
  location = c("all", "home", "school", "work", "other")
)

Arguments

countries

A character string or a vector of character containing the names of the countries for which to return contact data

location

Character. One of "all", "home", "school", "work" or "other".

Value

A data.frame (in long format) with 4 columns:

country: the country name

age_from: the age group of individual

age_to: the age group of contact

contact: the intensity of contact

References

Kiesha Prem, Alex R. Cook, Mark Jit, Projecting social contact matrices in 152 countries using contact surveys and demographic data, PLoS Comp. Biol. (2017), doi: 10.1371/journal.pcbi.1005697

Examples

contact_df_countries(c("Austria", "Belgium"), location = "all")
#>     country age_from age_to      contact
#> 1   Austria    00_05  00_05 1.327180e+00
#> 2   Austria    05_10  00_05 4.823092e-01
#> 3   Austria    10_15  00_05 1.745796e-01
#> 4   Austria    15_20  00_05 9.409321e-02
#> 5   Austria    20_25  00_05 1.543110e-01
#> 6   Austria    25_30  00_05 4.065126e-01
#> 7   Austria    30_35  00_05 5.776247e-01
#> 8   Austria    35_40  00_05 5.676579e-01
#> 9   Austria    40_45  00_05 3.407089e-01
#> 10  Austria    45_50  00_05 1.973056e-01
#> 11  Austria    50_55  00_05 2.000799e-01
#> 12  Austria    55_60  00_05 3.355600e-01
#> 13  Austria    60_65  00_05 3.449954e-01
#> 14  Austria    65_70  00_05 2.443926e-01
#> 15  Austria    70_75  00_05 1.453220e-01
#> 16  Austria    75_80  00_05 2.211635e-01
#> 17  Austria    00_05  05_10 5.763732e-01
#> 18  Austria    05_10  05_10 5.858879e+00
#> 19  Austria    10_15  05_10 1.637193e+00
#> 20  Austria    15_20  05_10 2.967136e-01
#> 21  Austria    20_25  05_10 1.675495e-01
#> 22  Austria    25_30  05_10 2.607602e-01
#> 23  Austria    30_35  05_10 8.208401e-01
#> 24  Austria    35_40  05_10 9.486506e-01
#> 25  Austria    40_45  05_10 7.866678e-01
#> 26  Austria    45_50  05_10 6.115721e-01
#> 27  Austria    50_55  05_10 6.967842e-01
#> 28  Austria    55_60  05_10 6.529761e-01
#> 29  Austria    60_65  05_10 3.358587e-01
#> 30  Austria    65_70  05_10 3.732483e-01
#> 31  Austria    70_75  05_10 3.980436e-01
#> 32  Austria    75_80  05_10 2.963571e-01
#> 33  Austria    00_05  10_15 2.566324e-01
#> 34  Austria    05_10  10_15 8.694872e-01
#> 35  Austria    10_15  10_15 7.936915e+00
#> 36  Austria    15_20  10_15 2.984048e+00
#> 37  Austria    20_25  10_15 2.790625e-01
#> 38  Austria    25_30  10_15 2.043004e-01
#> 39  Austria    30_35  10_15 5.902769e-01
#> 40  Austria    35_40  10_15 7.769115e-01
#> 41  Austria    40_45  10_15 1.052309e+00
#> 42  Austria    45_50  10_15 7.682170e-01
#> 43  Austria    50_55  10_15 1.044065e+00
#> 44  Austria    55_60  10_15 5.958005e-01
#> 45  Austria    60_65  10_15 2.372172e-01
#> 46  Austria    65_70  10_15 3.043497e-01
#> 47  Austria    70_75  10_15 4.138197e-01
#> 48  Austria    75_80  10_15 4.298191e-01
#> 49  Austria    00_05  15_20 1.591361e-01
#> 50  Austria    05_10  15_20 2.431130e-01
#> 51  Austria    10_15  15_20 7.960384e-01
#> 52  Austria    15_20  15_20 1.212499e+01
#> 53  Austria    20_25  15_20 2.412488e+00
#> 54  Austria    25_30  15_20 8.865504e-01
#> 55  Austria    30_35  15_20 5.123513e-01
#> 56  Austria    35_40  15_20 7.788524e-01
#> 57  Austria    40_45  15_20 1.324308e+00
#> 58  Austria    45_50  15_20 1.949179e+00
#> 59  Austria    50_55  15_20 1.521043e+00
#> 60  Austria    55_60  15_20 8.012565e-01
#> 61  Austria    60_65  15_20 3.712721e-01
#> 62  Austria    65_70  15_20 1.828274e-01
#> 63  Austria    70_75  15_20 4.314168e-01
#> 64  Austria    75_80  15_20 3.555489e-01
#> 65  Austria    00_05  20_25 2.562061e-01
#> 66  Austria    05_10  20_25 1.484631e-01
#> 67  Austria    10_15  20_25 2.728859e-01
#> 68  Austria    15_20  20_25 1.486105e+00
#> 69  Austria    20_25  20_25 4.012172e+00
#> 70  Austria    25_30  20_25 2.051951e+00
#> 71  Austria    30_35  20_25 1.042723e+00
#> 72  Austria    35_40  20_25 7.694590e-01
#> 73  Austria    40_45  20_25 9.808326e-01
#> 74  Austria    45_50  20_25 9.979113e-01
#> 75  Austria    50_55  20_25 1.061741e+00
#> 76  Austria    55_60  20_25 5.876676e-01
#> 77  Austria    60_65  20_25 3.533933e-01
#> 78  Austria    65_70  20_25 2.634855e-01
#> 79  Austria    70_75  20_25 2.164153e-01
#> 80  Austria    75_80  20_25 1.555979e-01
#> 81  Austria    00_05  25_30 4.785609e-01
#> 82  Austria    05_10  25_30 3.635834e-01
#> 83  Austria    10_15  25_30 2.421403e-01
#> 84  Austria    15_20  25_30 7.577318e-01
#> 85  Austria    20_25  25_30 1.761976e+00
#> 86  Austria    25_30  25_30 3.590935e+00
#> 87  Austria    30_35  25_30 1.777758e+00
#> 88  Austria    35_40  25_30 1.426896e+00
#> 89  Austria    40_45  25_30 1.330700e+00
#> 90  Austria    45_50  25_30 1.092148e+00
#> 91  Austria    50_55  25_30 1.312478e+00
#> 92  Austria    55_60  25_30 9.523424e-01
#> 93  Austria    60_65  25_30 5.496512e-01
#> 94  Austria    65_70  25_30 3.994821e-01
#> 95  Austria    70_75  25_30 3.681591e-01
#> 96  Austria    75_80  25_30 1.832528e-01
#> 97  Austria    00_05  30_35 7.405927e-01
#> 98  Austria    05_10  30_35 6.576091e-01
#> 99  Austria    10_15  30_35 3.937454e-01
#> 100 Austria    15_20  30_35 6.268761e-01
#> 101 Austria    20_25  30_35 1.203246e+00
#> 102 Austria    25_30  30_35 1.848242e+00
#> 103 Austria    30_35  30_35 2.992273e+00
#> 104 Austria    35_40  30_35 1.778268e+00
#> 105 Austria    40_45  30_35 1.702329e+00
#> 106 Austria    45_50  30_35 1.351557e+00
#> 107 Austria    50_55  30_35 1.179968e+00
#> 108 Austria    55_60  30_35 9.287872e-01
#> 109 Austria    60_65  30_35 6.938067e-01
#> 110 Austria    65_70  30_35 6.292624e-01
#> 111 Austria    70_75  30_35 3.967249e-01
#> 112 Austria    75_80  30_35 3.031298e-01
#> 113 Austria    00_05  35_40 7.328849e-01
#> 114 Austria    05_10  35_40 8.245483e-01
#> 115 Austria    10_15  35_40 6.688421e-01
#> 116 Austria    15_20  35_40 8.311498e-01
#> 117 Austria    20_25  35_40 1.142266e+00
#> 118 Austria    25_30  35_40 1.495496e+00
#> 119 Austria    30_35  35_40 1.942573e+00
#> 120 Austria    35_40  35_40 3.109081e+00
#> 121 Austria    40_45  35_40 1.984863e+00
#> 122 Austria    45_50  35_40 1.555159e+00
#> 123 Austria    50_55  35_40 1.181520e+00
#> 124 Austria    55_60  35_40 7.012638e-01
#> 125 Austria    60_65  35_40 6.946345e-01
#> 126 Austria    65_70  35_40 6.199822e-01
#> 127 Austria    70_75  35_40 6.803975e-01
#> 128 Austria    75_80  35_40 3.717528e-01
#> 129 Austria    00_05  40_45 4.694449e-01
#> 130 Austria    05_10  40_45 8.423361e-01
#> 131 Austria    10_15  40_45 1.053835e+00
#> 132 Austria    15_20  40_45 1.131942e+00
#> 133 Austria    20_25  40_45 1.015669e+00
#> 134 Austria    25_30  40_45 1.337537e+00
#> 135 Austria    30_35  40_45 1.548324e+00
#> 136 Austria    35_40  40_45 2.201498e+00
#> 137 Austria    40_45  40_45 3.229027e+00
#> 138 Austria    45_50  40_45 1.764583e+00
#> 139 Austria    50_55  40_45 1.645231e+00
#> 140 Austria    55_60  40_45 8.278159e-01
#> 141 Austria    60_65  40_45 5.809907e-01
#> 142 Austria    65_70  40_45 6.178417e-01
#> 143 Austria    70_75  40_45 8.194341e-01
#> 144 Austria    75_80  40_45 4.644301e-01
#> 145 Austria    00_05  45_50 2.335089e-01
#> 146 Austria    05_10  45_50 3.562218e-01
#> 147 Austria    10_15  45_50 5.840006e-01
#> 148 Austria    15_20  45_50 1.166779e+00
#> 149 Austria    20_25  45_50 1.271287e+00
#> 150 Austria    25_30  45_50 1.122469e+00
#> 151 Austria    30_35  45_50 1.214565e+00
#> 152 Austria    35_40  45_50 1.419468e+00
#> 153 Austria    40_45  45_50 1.854626e+00
#> 154 Austria    45_50  45_50 2.484463e+00
#> 155 Austria    50_55  45_50 1.854294e+00
#> 156 Austria    55_60  45_50 7.328722e-01
#> 157 Austria    60_65  45_50 4.942828e-01
#> 158 Austria    65_70  45_50 4.125454e-01
#> 159 Austria    70_75  45_50 6.693161e-01
#> 160 Austria    75_80  45_50 5.276658e-01
#> 161 Austria    00_05  50_55 2.355815e-01
#> 162 Austria    05_10  50_55 2.152172e-01
#> 163 Austria    10_15  50_55 3.070328e-01
#> 164 Austria    15_20  50_55 5.854959e-01
#> 165 Austria    20_25  50_55 8.344682e-01
#> 166 Austria    25_30  50_55 1.015105e+00
#> 167 Austria    30_35  50_55 8.287241e-01
#> 168 Austria    35_40  50_55 8.902870e-01
#> 169 Austria    40_45  50_55 1.138042e+00
#> 170 Austria    45_50  50_55 1.196648e+00
#> 171 Austria    50_55  50_55 1.947062e+00
#> 172 Austria    55_60  50_55 9.131883e-01
#> 173 Austria    60_65  50_55 4.575881e-01
#> 174 Austria    65_70  50_55 4.209265e-01
#> 175 Austria    70_75  50_55 5.008693e-01
#> 176 Austria    75_80  50_55 5.189642e-01
#> 177 Austria    00_05  55_60 1.816694e-01
#> 178 Austria    05_10  55_60 1.554733e-01
#> 179 Austria    10_15  55_60 1.325051e-01
#> 180 Austria    15_20  55_60 1.803664e-01
#> 181 Austria    20_25  55_60 3.428977e-01
#> 182 Austria    25_30  55_60 4.425755e-01
#> 183 Austria    30_35  55_60 4.336405e-01
#> 184 Austria    35_40  55_60 3.433649e-01
#> 185 Austria    40_45  55_60 2.860845e-01
#> 186 Austria    45_50  55_60 3.734521e-01
#> 187 Austria    50_55  55_60 6.678996e-01
#> 188 Austria    55_60  55_60 1.311941e+00
#> 189 Austria    60_65  55_60 6.261823e-01
#> 190 Austria    65_70  55_60 5.126237e-01
#> 191 Austria    70_75  55_60 4.271502e-01
#> 192 Austria    75_80  55_60 3.201911e-01
#> 193 Austria    00_05  60_65 1.376162e-01
#> 194 Austria    05_10  60_65 1.339667e-01
#> 195 Austria    10_15  60_65 7.967664e-02
#> 196 Austria    15_20  60_65 7.253009e-02
#> 197 Austria    20_25  60_65 9.722288e-02
#> 198 Austria    25_30  60_65 1.610526e-01
#> 199 Austria    30_35  60_65 2.292114e-01
#> 200 Austria    35_40  60_65 2.561783e-01
#> 201 Austria    40_45  60_65 2.067888e-01
#> 202 Austria    45_50  60_65 1.793907e-01
#> 203 Austria    50_55  60_65 2.821888e-01
#> 204 Austria    55_60  60_65 4.720481e-01
#> 205 Austria    60_65  60_65 1.168928e+00
#> 206 Austria    65_70  60_65 5.761567e-01
#> 207 Austria    70_75  60_65 7.929362e-01
#> 208 Austria    75_80  60_65 2.447676e-01
#> 209 Austria    00_05  65_70 1.181146e-01
#> 210 Austria    05_10  65_70 1.061203e-01
#> 211 Austria    10_15  65_70 8.831880e-02
#> 212 Austria    15_20  65_70 5.867875e-02
#> 213 Austria    20_25  65_70 5.500634e-02
#> 214 Austria    25_30  65_70 7.210570e-02
#> 215 Austria    30_35  65_70 1.151971e-01
#> 216 Austria    35_40  65_70 1.852636e-01
#> 217 Austria    40_45  65_70 1.563289e-01
#> 218 Austria    45_50  65_70 1.213620e-01
#> 219 Austria    50_55  65_70 1.548169e-01
#> 220 Austria    55_60  65_70 2.455473e-01
#> 221 Austria    60_65  65_70 4.577683e-01
#> 222 Austria    65_70  65_70 1.147893e+00
#> 223 Austria    70_75  65_70 8.453641e-01
#> 224 Austria    75_80  65_70 3.644550e-01
#> 225 Austria    00_05  70_75 8.797816e-02
#> 226 Austria    05_10  70_75 6.082912e-02
#> 227 Austria    10_15  70_75 7.928980e-02
#> 228 Austria    15_20  70_75 4.237846e-02
#> 229 Austria    20_25  70_75 7.285139e-02
#> 230 Austria    25_30  70_75 4.366647e-02
#> 231 Austria    30_35  70_75 6.993954e-02
#> 232 Austria    35_40  70_75 1.277451e-01
#> 233 Austria    40_45  70_75 1.322671e-01
#> 234 Austria    45_50  70_75 1.314849e-01
#> 235 Austria    50_55  70_75 1.292512e-01
#> 236 Austria    55_60  70_75 1.477166e-01
#> 237 Austria    60_65  70_75 3.081319e-01
#> 238 Austria    65_70  70_75 3.614710e-01
#> 239 Austria    70_75  70_75 1.342478e+00
#> 240 Austria    75_80  70_75 3.716226e-01
#> 241 Austria    00_05  75_80 4.054908e-02
#> 242 Austria    05_10  75_80 4.174674e-02
#> 243 Austria    10_15  75_80 5.787181e-02
#> 244 Austria    15_20  75_80 2.669395e-02
#> 245 Austria    20_25  75_80 5.177944e-02
#> 246 Austria    25_30  75_80 2.978390e-02
#> 247 Austria    30_35  75_80 5.415282e-02
#> 248 Austria    35_40  75_80 4.269159e-02
#> 249 Austria    40_45  75_80 5.448594e-02
#> 250 Austria    45_50  75_80 1.039120e-01
#> 251 Austria    50_55  75_80 1.054921e-01
#> 252 Austria    55_60  75_80 9.373601e-02
#> 253 Austria    60_65  75_80 1.085559e-01
#> 254 Austria    65_70  75_80 1.456167e-01
#> 255 Austria    70_75  75_80 3.675267e-01
#> 256 Austria    75_80  75_80 5.117988e-01
#> 257 Belgium    00_05  00_05 3.433355e+00
#> 258 Belgium    05_10  00_05 9.187387e-01
#> 259 Belgium    10_15  00_05 3.010123e-01
#> 260 Belgium    15_20  00_05 2.509607e-01
#> 261 Belgium    20_25  00_05 5.542132e-02
#> 262 Belgium    25_30  00_05 5.456022e-01
#> 263 Belgium    30_35  00_05 6.744262e-01
#> 264 Belgium    35_40  00_05 8.503682e-01
#> 265 Belgium    40_45  00_05 3.025879e-01
#> 266 Belgium    45_50  00_05 1.947534e-02
#> 267 Belgium    50_55  00_05 1.800766e-01
#> 268 Belgium    55_60  00_05 2.999669e-01
#> 269 Belgium    60_65  00_05 7.362493e-02
#> 270 Belgium    65_70  00_05 3.891347e-02
#> 271 Belgium    70_75  00_05 2.466115e-02
#> 272 Belgium    75_80  00_05 1.419924e-60
#> 273 Belgium    00_05  05_10 7.081745e-01
#> 274 Belgium    05_10  05_10 5.471858e+00
#> 275 Belgium    10_15  05_10 9.766143e-01
#> 276 Belgium    15_20  05_10 5.881658e-01
#> 277 Belgium    20_25  05_10 4.782861e-01
#> 278 Belgium    25_30  05_10 6.105946e-02
#> 279 Belgium    30_35  05_10 7.668862e-01
#> 280 Belgium    35_40  05_10 7.761638e-01
#> 281 Belgium    40_45  05_10 2.439954e-01
#> 282 Belgium    45_50  05_10 2.721435e-02
#> 283 Belgium    50_55  05_10 3.668674e-01
#> 284 Belgium    55_60  05_10 9.682446e-02
#> 285 Belgium    60_65  05_10 1.669788e-02
#> 286 Belgium    65_70  05_10 1.744632e-01
#> 287 Belgium    70_75  05_10 1.535370e-01
#> 288 Belgium    75_80  05_10 3.306408e-52
#> 289 Belgium    00_05  10_15 5.011964e-01
#> 290 Belgium    05_10  10_15 7.718200e-01
#> 291 Belgium    10_15  10_15 8.814612e+00
#> 292 Belgium    15_20  10_15 1.238135e+00
#> 293 Belgium    20_25  10_15 5.042784e-01
#> 294 Belgium    25_30  10_15 1.531528e-02
#> 295 Belgium    30_35  10_15 1.847164e-01
#> 296 Belgium    35_40  10_15 7.306811e-01
#> 297 Belgium    40_45  10_15 8.737648e-01
#> 298 Belgium    45_50  10_15 3.740025e-01
#> 299 Belgium    50_55  10_15 2.630297e-01
#> 300 Belgium    55_60  10_15 1.468197e-01
#> 301 Belgium    60_65  10_15 5.749550e-02
#> 302 Belgium    65_70  10_15 4.004636e-02
#> 303 Belgium    70_75  10_15 2.021694e-01
#> 304 Belgium    75_80  10_15 1.111683e-45
#> 305 Belgium    00_05  15_20 1.336298e-01
#> 306 Belgium    05_10  15_20 2.012956e-01
#> 307 Belgium    10_15  15_20 8.774924e-01
#> 308 Belgium    15_20  15_20 7.588225e+00
#> 309 Belgium    20_25  15_20 2.202155e+00
#> 310 Belgium    25_30  15_20 4.696600e-01
#> 311 Belgium    30_35  15_20 4.834487e-01
#> 312 Belgium    35_40  15_20 3.693035e-01
#> 313 Belgium    40_45  15_20 5.485267e-01
#> 314 Belgium    45_50  15_20 9.079198e-01
#> 315 Belgium    50_55  15_20 3.898366e-01
#> 316 Belgium    55_60  15_20 1.840471e-01
#> 317 Belgium    60_65  15_20 8.457465e-02
#> 318 Belgium    65_70  15_20 4.434731e-02
#> 319 Belgium    70_75  15_20 2.657793e-01
#> 320 Belgium    75_80  15_20 1.217259e-42
#> 321 Belgium    00_05  20_25 2.955514e-01
#> 322 Belgium    05_10  20_25 1.267253e-01
#> 323 Belgium    10_15  20_25 5.798680e-01
#> 324 Belgium    15_20  20_25 1.655134e+00
#> 325 Belgium    20_25  20_25 6.248385e+00
#> 326 Belgium    25_30  20_25 1.271145e+00
#> 327 Belgium    30_35  20_25 9.491313e-01
#> 328 Belgium    35_40  20_25 2.406077e-01
#> 329 Belgium    40_45  20_25 7.098732e-01
#> 330 Belgium    45_50  20_25 8.148212e-01
#> 331 Belgium    50_55  20_25 1.133704e+00
#> 332 Belgium    55_60  20_25 6.354115e-01
#> 333 Belgium    60_65  20_25 2.535381e-01
#> 334 Belgium    65_70  20_25 9.476837e-02
#> 335 Belgium    70_75  20_25 7.246614e-02
#> 336 Belgium    75_80  20_25 3.287066e-47
#> 337 Belgium    00_05  25_30 1.011272e+00
#> 338 Belgium    05_10  25_30 4.033151e-01
#> 339 Belgium    10_15  25_30 2.545030e-01
#> 340 Belgium    15_20  25_30 6.821279e-01
#> 341 Belgium    20_25  25_30 2.026125e+00
#> 342 Belgium    25_30  25_30 4.053200e+00
#> 343 Belgium    30_35  25_30 1.606787e+00
#> 344 Belgium    35_40  25_30 7.999242e-01
#> 345 Belgium    40_45  25_30 1.213600e+00
#> 346 Belgium    45_50  25_30 1.212942e+00
#> 347 Belgium    50_55  25_30 1.668529e+00
#> 348 Belgium    55_60  25_30 9.122716e-01
#> 349 Belgium    60_65  25_30 4.940597e-01
#> 350 Belgium    65_70  25_30 1.822151e-01
#> 351 Belgium    70_75  25_30 2.409876e-01
#> 352 Belgium    75_80  25_30 5.463250e-02
#> 353 Belgium    00_05  30_35 1.562276e+00
#> 354 Belgium    05_10  30_35 1.158775e+00
#> 355 Belgium    10_15  30_35 4.186921e-01
#> 356 Belgium    15_20  30_35 2.991406e-01
#> 357 Belgium    20_25  30_35 1.697659e+00
#> 358 Belgium    25_30  30_35 1.838385e+00
#> 359 Belgium    30_35  30_35 3.881850e+00
#> 360 Belgium    35_40  30_35 1.527024e+00
#> 361 Belgium    40_45  30_35 1.162387e+00
#> 362 Belgium    45_50  30_35 1.208749e+00
#> 363 Belgium    50_55  30_35 9.951900e-01
#> 364 Belgium    55_60  30_35 1.241810e+00
#> 365 Belgium    60_65  30_35 8.310562e-01
#> 366 Belgium    65_70  30_35 4.868882e-01
#> 367 Belgium    70_75  30_35 2.349565e-01
#> 368 Belgium    75_80  30_35 2.140319e-01
#> 369 Belgium    00_05  35_40 8.170258e-01
#> 370 Belgium    05_10  35_40 8.945993e-01
#> 371 Belgium    10_15  35_40 1.180393e+00
#> 372 Belgium    15_20  35_40 5.263672e-01
#> 373 Belgium    20_25  35_40 6.877021e-01
#> 374 Belgium    25_30  35_40 1.273790e+00
#> 375 Belgium    30_35  35_40 2.582737e+00
#> 376 Belgium    35_40  35_40 2.824511e+00
#> 377 Belgium    40_45  35_40 1.474908e+00
#> 378 Belgium    45_50  35_40 2.086639e+00
#> 379 Belgium    50_55  35_40 1.040404e+00
#> 380 Belgium    55_60  35_40 8.932932e-01
#> 381 Belgium    60_65  35_40 7.307260e-01
#> 382 Belgium    65_70  35_40 5.812430e-01
#> 383 Belgium    70_75  35_40 5.012109e-01
#> 384 Belgium    75_80  35_40 1.890592e-02
#> 385 Belgium    00_05  40_45 1.861119e-01
#> 386 Belgium    05_10  40_45 6.539895e-01
#> 387 Belgium    10_15  40_45 1.212396e+00
#> 388 Belgium    15_20  40_45 8.256399e-01
#> 389 Belgium    20_25  40_45 1.473976e+00
#> 390 Belgium    25_30  40_45 1.172393e+00
#> 391 Belgium    30_35  40_45 1.390617e+00
#> 392 Belgium    35_40  40_45 1.869020e+00
#> 393 Belgium    40_45  40_45 2.824161e+00
#> 394 Belgium    45_50  40_45 1.269894e+00
#> 395 Belgium    50_55  40_45 1.543598e+00
#> 396 Belgium    55_60  40_45 1.053617e+00
#> 397 Belgium    60_65  40_45 6.567880e-01
#> 398 Belgium    65_70  40_45 7.005581e-01
#> 399 Belgium    70_75  40_45 6.048176e-01
#> 400 Belgium    75_80  40_45 5.998784e-02
#> 401 Belgium    00_05  45_50 1.491145e-01
#> 402 Belgium    05_10  45_50 2.828605e-01
#> 403 Belgium    10_15  45_50 7.159357e-01
#> 404 Belgium    15_20  45_50 1.208740e+00
#> 405 Belgium    20_25  45_50 1.529449e+00
#> 406 Belgium    25_30  45_50 7.612885e-01
#> 407 Belgium    30_35  45_50 1.124795e+00
#> 408 Belgium    35_40  45_50 7.739052e-01
#> 409 Belgium    40_45  45_50 1.244328e+00
#> 410 Belgium    45_50  45_50 3.825359e+00
#> 411 Belgium    50_55  45_50 1.307075e+00
#> 412 Belgium    55_60  45_50 8.579458e-01
#> 413 Belgium    60_65  45_50 5.351167e-01
#> 414 Belgium    65_70  45_50 4.413039e-01
#> 415 Belgium    70_75  45_50 5.512242e-01
#> 416 Belgium    75_80  45_50 5.257872e-02
#> 417 Belgium    00_05  50_55 4.290754e-01
#> 418 Belgium    05_10  50_55 1.906483e-01
#> 419 Belgium    10_15  50_55 1.482475e-01
#> 420 Belgium    15_20  50_55 3.317466e-01
#> 421 Belgium    20_25  50_55 8.550549e-01
#> 422 Belgium    25_30  50_55 8.461579e-01
#> 423 Belgium    30_35  50_55 5.465913e-01
#> 424 Belgium    35_40  50_55 5.518361e-01
#> 425 Belgium    40_45  50_55 7.841036e-01
#> 426 Belgium    45_50  50_55 8.465074e-01
#> 427 Belgium    50_55  50_55 2.233946e+00
#> 428 Belgium    55_60  50_55 1.241051e+00
#> 429 Belgium    60_65  50_55 3.764918e-01
#> 430 Belgium    65_70  50_55 9.434179e-02
#> 431 Belgium    70_75  50_55 6.905890e-01
#> 432 Belgium    75_80  50_55 2.818647e-01
#> 433 Belgium    00_05  55_60 3.180308e-01
#> 434 Belgium    05_10  55_60 2.536995e-01
#> 435 Belgium    10_15  55_60 2.291342e-01
#> 436 Belgium    15_20  55_60 8.499961e-02
#> 437 Belgium    20_25  55_60 2.918044e-01
#> 438 Belgium    25_30  55_60 1.997267e-01
#> 439 Belgium    30_35  55_60 5.457430e-01
#> 440 Belgium    35_40  55_60 2.942780e-01
#> 441 Belgium    40_45  55_60 8.810123e-02
#> 442 Belgium    45_50  55_60 4.631469e-01
#> 443 Belgium    50_55  55_60 9.198801e-01
#> 444 Belgium    55_60  55_60 1.733285e+00
#> 445 Belgium    60_65  55_60 8.000309e-01
#> 446 Belgium    65_70  55_60 3.676999e-01
#> 447 Belgium    70_75  55_60 2.411997e-01
#> 448 Belgium    75_80  55_60 1.612294e-01
#> 449 Belgium    00_05  60_65 1.553783e-01
#> 450 Belgium    05_10  60_65 2.001432e-01
#> 451 Belgium    10_15  60_65 2.713757e-01
#> 452 Belgium    15_20  60_65 7.744645e-02
#> 453 Belgium    20_25  60_65 1.283884e-01
#> 454 Belgium    25_30  60_65 2.904546e-02
#> 455 Belgium    30_35  60_65 1.405671e-01
#> 456 Belgium    35_40  60_65 1.870542e-01
#> 457 Belgium    40_45  60_65 1.420546e-01
#> 458 Belgium    45_50  60_65 7.295081e-02
#> 459 Belgium    50_55  60_65 5.946344e-01
#> 460 Belgium    55_60  60_65 5.479298e-01
#> 461 Belgium    60_65  60_65 2.384264e+00
#> 462 Belgium    65_70  60_65 4.698754e-01
#> 463 Belgium    70_75  60_65 1.099401e+00
#> 464 Belgium    75_80  60_65 9.649168e-02
#> 465 Belgium    00_05  65_70 1.473301e-01
#> 466 Belgium    05_10  65_70 5.622217e-02
#> 467 Belgium    10_15  65_70 1.153940e-01
#> 468 Belgium    15_20  65_70 9.148703e-02
#> 469 Belgium    20_25  65_70 8.223947e-02
#> 470 Belgium    25_30  65_70 4.740496e-02
#> 471 Belgium    30_35  65_70 9.411408e-02
#> 472 Belgium    35_40  65_70 1.094134e-01
#> 473 Belgium    40_45  65_70 1.169574e-01
#> 474 Belgium    45_50  65_70 5.600277e-02
#> 475 Belgium    50_55  65_70 2.135735e-01
#> 476 Belgium    55_60  65_70 1.747342e-01
#> 477 Belgium    60_65  65_70 8.718327e-01
#> 478 Belgium    65_70  65_70 1.502259e+00
#> 479 Belgium    70_75  65_70 5.418966e-01
#> 480 Belgium    75_80  65_70 3.445393e-01
#> 481 Belgium    00_05  70_75 5.971800e-02
#> 482 Belgium    05_10  70_75 6.668890e-02
#> 483 Belgium    10_15  70_75 7.132689e-02
#> 484 Belgium    15_20  70_75 4.687557e-03
#> 485 Belgium    20_25  70_75 1.047583e-01
#> 486 Belgium    25_30  70_75 1.309121e-02
#> 487 Belgium    30_35  70_75 1.236489e-02
#> 488 Belgium    35_40  70_75 1.604198e-01
#> 489 Belgium    40_45  70_75 1.481748e-01
#> 490 Belgium    45_50  70_75 4.912546e-02
#> 491 Belgium    50_55  70_75 1.862894e-01
#> 492 Belgium    55_60  70_75 7.316012e-02
#> 493 Belgium    60_65  70_75 7.223738e-01
#> 494 Belgium    65_70  70_75 3.916101e-01
#> 495 Belgium    70_75  70_75 1.156188e+00
#> 496 Belgium    75_80  70_75 6.606435e-01
#> 497 Belgium    00_05  75_80 8.000076e-02
#> 498 Belgium    05_10  75_80 5.430751e-02
#> 499 Belgium    10_15  75_80 6.910402e-02
#> 500 Belgium    15_20  75_80 5.200677e-03
#> 501 Belgium    20_25  75_80 1.090817e-01
#> 502 Belgium    25_30  75_80 5.645883e-02
#> 503 Belgium    30_35  75_80 1.645378e-01
#> 504 Belgium    35_40  75_80 1.268286e-02
#> 505 Belgium    40_45  75_80 4.146498e-02
#> 506 Belgium    45_50  75_80 8.183890e-03
#> 507 Belgium    50_55  75_80 2.280204e-01
#> 508 Belgium    55_60  75_80 1.822894e-01
#> 509 Belgium    60_65  75_80 2.930788e-01
#> 510 Belgium    65_70  75_80 2.515207e-01
#> 511 Belgium    70_75  75_80 5.135013e-01
#> 512 Belgium    75_80  75_80 4.871102e-01