Get contact data matrix for a specific country

Get contact data matrix for a specific country

contact_matrix(country, location = c("all", "home", "school", "work", "other"))

Arguments

country

Character. The name of the country for which you want contact data.

location

Character. One of "all", "home", "school", "work" or "other".

Value

A square (16 by 16) matrix containing the contact data between the different age classes for a given country.

References

Kiesha Prem, Alex R. Cook, Mark Jit, Projecting social contact matrices in 152 countries using contact surveys and demographic data, PLoS Comp. Biol. (2017), doi: 10.1371/journal.pcbi.1005697

Examples

contact_matrix("France", location = "all")
#>           00_05     05_10     10_15     15_20     20_25     25_30     30_35
#> 00_05 3.8049208 1.1064463 0.4119145 0.2553693 0.3417530 0.7532100 1.2090488
#> 05_10 1.0620240 5.0325631 1.0108210 0.2729936 0.1637277 0.4151037 0.9207158
#> 10_15 0.2383228 1.5384390 6.9859632 0.8459108 0.2870553 0.2443827 0.4032533
#> 15_20 0.1242916 0.3084664 2.3013378 7.8316731 1.3599526 0.6511973 0.5309225
#> 20_25 0.1997494 0.1682122 0.2555462 2.1654802 3.9337904 1.7076407 1.1677744
#> 25_30 0.5331867 0.2486300 0.1435870 0.7309595 1.9436259 3.4581082 1.7769048
#> 30_35 0.7222459 0.8480973 0.5299038 0.4228065 1.0050076 1.7204472 2.9192993
#> 35_40 0.7037954 1.0721441 0.8366342 0.6832771 0.7646973 1.4071985 1.7597999
#> 40_45 0.3095332 0.6566442 0.9827524 1.1093668 0.9235031 1.2518923 1.5914371
#> 45_50 0.4058094 0.4671793 0.6117632 1.5367229 0.9454052 0.9938984 1.2592405
#> 50_55 0.2517178 0.5947777 0.8288772 1.2016733 1.0188128 1.2906068 1.1792561
#> 55_60 0.5424359 0.6711899 0.5736708 0.7492978 0.6135626 0.9317515 0.9175372
#> 60_65 0.3961294 0.3629084 0.2566179 0.3614150 0.3387789 0.5080470 0.6393848
#> 65_70 0.1980189 0.3134375 0.2568708 0.1574851 0.2121279 0.3071351 0.4501078
#> 70_75 0.1052495 0.2944891 0.3077932 0.3553412 0.1558914 0.2445032 0.2530902
#> 75_80 0.2435584 0.3171220 0.4522256 0.3598813 0.1542999 0.1882229 0.3050821
#>           35_40     40_45     45_50     50_55     55_60      60_65      65_70
#> 00_05 1.0419960 0.4955829 0.3104594 0.3070544 0.2722105 0.18633754 0.12804520
#> 05_10 1.1384301 0.8451808 0.3314044 0.2073083 0.1854381 0.17691472 0.11164899
#> 10_15 0.8269250 1.0806856 0.5704767 0.2657533 0.1466135 0.10152928 0.08817052
#> 15_20 0.7621321 1.0248384 1.0489743 0.4785021 0.1639009 0.07173157 0.05282605
#> 20_25 1.1331411 0.9700452 1.2702652 0.8133999 0.3272061 0.08382776 0.04576510
#> 25_30 1.4707826 1.2626267 0.9969784 0.9610877 0.3574627 0.11253397 0.05165248
#> 30_35 1.8937794 1.4574822 1.1351731 0.8100567 0.4053920 0.18410812 0.08541583
#> 35_40 3.2108263 2.0977977 1.3443554 0.9195016 0.3641096 0.25969059 0.15240004
#> 40_45 1.8766171 2.9541130 1.6818322 1.1128901 0.2732150 0.18353558 0.11084809
#> 45_50 1.4654930 1.5731579 2.1964738 1.0863042 0.3406847 0.15129942 0.08636681
#> 50_55 1.2145400 1.5942381 1.7430635 1.9184501 0.6491669 0.26726729 0.11850585
#> 55_60 0.7479099 0.7927704 0.6808703 0.9229155 1.4619218 0.51294863 0.21825949
#> 60_65 0.6940436 0.5234400 0.4162769 0.4414520 0.6750623 1.40597284 0.42649139
#> 65_70 0.4543158 0.4253709 0.3023152 0.3242779 0.4404185 0.51803705 1.00304348
#> 70_75 0.4770518 0.5149822 0.4340009 0.3392186 0.3381533 0.65997723 0.61920693
#> 75_80 0.3824259 0.4514450 0.4428716 0.5388023 0.3440813 0.27980028 0.36715407
#>            70_75      75_80
#> 00_05 0.08780229 0.05386407
#> 05_10 0.05825985 0.05311071
#> 10_15 0.06970996 0.06966683
#> 15_20 0.03453706 0.02613884
#> 20_25 0.05447291 0.05237283
#> 25_30 0.03057661 0.02361016
#> 30_35 0.04700946 0.05253904
#> 35_40 0.09805799 0.04637476
#> 40_45 0.08697388 0.04801201
#> 45_50 0.08310938 0.08187873
#> 50_55 0.08806304 0.09026068
#> 55_60 0.11994582 0.09467785
#> 60_65 0.26297656 0.12855394
#> 65_70 0.27590642 0.13898285
#> 70_75 0.97407644 0.32576427
#> 75_80 0.36754549 0.64291815

contact_matrix("Belgium", location = "school")
#>               00_05         05_10         10_15         15_20         20_25
#> 00_05  2.231562e+00  2.188485e-01  3.900767e-01  2.688331e-02  1.032931e-02
#> 05_10  2.472377e-01  3.117592e+00  2.057124e-01  1.607709e-02  9.661761e-19
#> 10_15  7.631087e-02  1.573915e-01  4.195039e+00  1.005322e-01  6.057941e-02
#> 15_20  1.667880e-01  1.188428e-01  5.047092e-01  4.047474e+00  6.130302e-02
#> 20_25  3.797864e-27  3.165459e-30  1.110084e-64  2.179770e-01  2.186010e-01
#> 25_30  5.860132e-52  3.296817e-44  1.510324e-61  8.547018e-39  9.412904e-34
#> 30_35  5.880712e-02  1.027104e-01  1.469999e-45  5.640773e-23  5.685250e-03
#> 35_40  4.086717e-01  1.705787e-16  1.546249e-01  1.873899e-51  3.260699e-02
#> 40_45  1.414248e-01  7.960342e-45  2.448759e-57  3.152303e-02  7.128404e-43
#> 45_50  5.537966e-63 9.859117e-104  1.576919e-49  4.952548e-02  1.134950e-38
#> 50_55  3.252134e-53 8.848808e-134  3.248688e-33  5.037806e-85 2.965335e-101
#> 55_60  2.937304e-02  2.816686e-02  7.384065e-02  2.832956e-02 9.790374e-127
#> 60_65  6.079235e-69  1.214646e-47  3.733487e-02  7.817605e-75 4.152412e-105
#> 65_70 1.740252e-141 9.780961e-109  2.514760e-99  5.718298e-89 2.861488e-147
#> 70_75 5.395875e-124 5.013824e-125 9.620625e-104 5.056621e-123 4.011300e-115
#> 75_80 1.641101e-116 6.770031e-123 1.936422e-116 1.190831e-119 4.543822e-119
#>               25_30        30_35        35_40        40_45        45_50
#> 00_05  1.738315e-01 2.889025e-01 7.389466e-02 4.453167e-02 4.535417e-14
#> 05_10  6.256750e-02 7.893416e-02 7.008973e-02 1.198746e-01 7.160435e-02
#> 10_15  5.682804e-02 8.798125e-03 7.864829e-02 6.385997e-02 1.542912e-01
#> 15_20  5.795102e-02 1.697930e-25 1.325180e-01 3.960846e-02 1.105025e-01
#> 20_25  1.046959e-02 2.013059e-28 3.447825e-22 2.661493e-02 1.153428e-02
#> 25_30  3.074052e-01 2.812532e-44 1.864332e-36 3.672584e-54 5.624769e-62
#> 30_35  8.176267e-03 3.597535e-02 5.285380e-02 5.494858e-02 8.736286e-03
#> 35_40  1.851403e-01 8.048927e-02 6.714555e-02 8.853332e-02 3.042403e-02
#> 40_45  4.827482e-02 8.277887e-02 1.984801e-02 9.134568e-02 2.011418e-02
#> 45_50  4.636087e-65 1.108271e-25 8.792324e-02 6.012818e-51 1.431795e-46
#> 50_55 5.525322e-133 8.965046e-43 2.650957e-02 2.715616e-02 2.235838e-43
#> 55_60 2.047690e-126 5.069021e-02 1.980646e-02 5.330642e-34 1.999003e-02
#> 60_65  1.724260e-81 1.789528e-78 1.359651e-01 3.079255e-02 2.328519e-35
#> 65_70 1.387484e-166 4.157155e-51 8.389592e-62 4.998088e-43 5.506727e-02
#> 70_75  2.139350e-97 7.074939e-87 1.661226e-02 1.674010e-02 4.565433e-02
#> 75_80 4.235386e-126 3.939033e-76 1.222147e-94 1.245041e-86 2.734155e-67
#>              50_55        55_60         60_65         65_70         70_75
#> 00_05 6.800606e-02 5.251640e-48  1.133523e-90 3.079103e-104  2.003412e-75
#> 05_10 4.988977e-02 1.683643e-02  6.127720e-99  2.056286e-03  1.979552e-03
#> 10_15 4.046869e-02 1.982978e-02  3.923117e-03  9.768538e-53  8.221526e-43
#> 15_20 5.163183e-02 2.154750e-02 1.101372e-102  5.472303e-80 7.378277e-103
#> 20_25 1.562146e-40 2.238559e-79  7.644916e-43 2.398178e-139 2.483573e-108
#> 25_30 7.031231e-84 3.013535e-61 5.939161e-126  4.614686e-02  1.654861e-93
#> 30_35 2.153146e-02 3.187020e-61  4.323460e-03  1.080235e-99  2.525521e-70
#> 35_40 4.654946e-31 8.013485e-03  8.639725e-79 1.496899e-107 9.645247e-135
#> 40_45 5.053899e-02 1.479312e-02  7.719070e-43  2.760307e-88 5.801369e-138
#> 45_50 1.009745e-60 1.585206e-93  1.558346e-02 5.605592e-134 2.724760e-108
#> 50_55 1.530534e-81 2.273165e-42  7.685372e-96  6.001953e-62 1.796864e-131
#> 55_60 4.992229e-02 4.560906e-84  1.106032e-02  6.447958e-87 8.829809e-111
#> 60_65 2.091794e-41 1.744472e-67 1.310905e-115 1.327183e-114 1.677496e-139
#> 65_70 5.154972e-51 1.628736e-87  1.082476e-89  2.923436e-83  1.683342e-64
#> 70_75 4.590242e-02 4.587703e-02  1.664027e-02  1.659724e-02  1.654680e-02
#> 75_80 2.462743e-65 3.462799e-68  1.218998e-76  8.387174e-76  1.051333e-68
#>               75_80
#> 00_05 9.377263e-107
#> 05_10  2.368802e-85
#> 10_15  2.315574e-03
#> 15_20 3.734503e-127
#> 20_25  5.385756e-97
#> 25_30  1.219396e-96
#> 30_35 1.878419e-137
#> 35_40 3.375347e-129
#> 40_45 9.617777e-117
#> 45_50 1.262355e-111
#> 50_55 4.112569e-130
#> 55_60 5.214644e-120
#> 60_65 6.424454e-111
#> 65_70  5.775052e-74
#> 70_75  4.622517e-02
#> 75_80  1.014157e-96