Get the list of countries included in the dataset

Get the list of countries included in the dataset

list_countries(
  geographic_setting = c("all", "rural", "urban"),
  data_source = c("2020", "2017")
)

Arguments

geographic_setting

Character. One of "all" (default), "rural", "urban"

data_source

Character. Either "202O" (default) or "2017"

Value

A character vector with the name of all countries included in the dataset

Note

This package uses the nomenclature from the countrycode package. If your names differ from the names used here, you should use countrycode as well to update them.

References

Kiesha Prem, Alex R. Cook, Mark Jit, Projecting social contact matrices in 152 countries using contact surveys and demographic data, PLoS Comp. Biol. (2017), doi:10.1371/journal.pcbi.1005697

Kiesha Prem, Kevin van Zandvoort, Petra Klepac, Rosalind M. Eggo, Nicholas G. Davies, CMMID COVID-19 Working Group, Alex R. Cook, Mark Jit, Projecting contact matrices in 177 geographical regions: An update and comparison with empirical data for the COVID-19 era, PLoS Comp. Biol. (2021), doi:10.1371/journal.pcbi.1009098 .

Examples

list_countries()
#>   [1] "Afghanistan"              "Angola"                  
#>   [3] "Albania"                  "United Arab Emirates"    
#>   [5] "Argentina"                "Armenia"                 
#>   [7] "Austria"                  "Azerbaijan"              
#>   [9] "Burundi"                  "Belgium"                 
#>  [11] "Benin"                    "Burkina Faso"            
#>  [13] "Bangladesh"               "Bulgaria"                
#>  [15] "Bahrain"                  "Bahamas"                 
#>  [17] "Bosnia"                   "Belarus"                 
#>  [19] "Belize"                   "Bolivia"                 
#>  [21] "Brazil"                   "Barbados"                
#>  [23] "Brunei"                   "Bhutan"                  
#>  [25] "Botswana"                 "Central African Republic"
#>  [27] "Canada"                   "Switzerland"             
#>  [29] "Chile"                    "China"                   
#>  [31] "Côte d’Ivoire"            "Cameroon"                
#>  [33] "Congo - Kinshasa"         "Congo - Brazzaville"     
#>  [35] "Colombia"                 "Comoros"                 
#>  [37] "Cape Verde"               "Costa Rica"              
#>  [39] "Cuba"                     "Cyprus"                  
#>  [41] "Czechia"                  "Germany"                 
#>  [43] "Djibouti"                 "Denmark"                 
#>  [45] "Dominican Republic"       "Algeria"                 
#>  [47] "Ecuador"                  "Egypt"                   
#>  [49] "Eritrea"                  "Spain"                   
#>  [51] "Estonia"                  "Ethiopia"                
#>  [53] "Finland"                  "Fiji"                    
#>  [55] "France"                   "Gabon"                   
#>  [57] "UK"                       "Georgia"                 
#>  [59] "Ghana"                    "Guinea"                  
#>  [61] "Gambia"                   "Guinea-Bissau"           
#>  [63] "Equatorial Guinea"        "Greece"                  
#>  [65] "Guatemala"                "Guyana"                  
#>  [67] "Hong Kong"                "Honduras"                
#>  [69] "Croatia"                  "Hungary"                 
#>  [71] "Indonesia"                "India"                   
#>  [73] "Ireland"                  "Iran"                    
#>  [75] "Iraq"                     "Iceland"                 
#>  [77] "Israel"                   "Italy"                   
#>  [79] "Jamaica"                  "Jordan"                  
#>  [81] "Kazakhstan"               "Kenya"                   
#>  [83] "Kyrgyzstan"               "Cambodia"                
#>  [85] "South Korea"              "Kuwait"                  
#>  [87] "Laos"                     "Liberia"                 
#>  [89] "Libya"                    "St. Lucia"               
#>  [91] "Sri Lanka"                "Lesotho"                 
#>  [93] "Lithuania"                "Luxembourg"              
#>  [95] "Latvia"                   "Macao"                   
#>  [97] "Morocco"                  "Moldova"                 
#>  [99] "Madagascar"               "Maldives"                
#> [101] "Mexico"                   "North Macedonia"         
#> [103] "Mali"                     "Malta"                   
#> [105] "Myanmar"                  "Montenegro"              
#> [107] "Mongolia"                 "Mozambique"              
#> [109] "Mauritania"               "Mauritius"               
#> [111] "Malawi"                   "Malaysia"                
#> [113] "Namibia"                  "Niger"                   
#> [115] "Nigeria"                  "Nicaragua"               
#> [117] "Netherlands"              "Norway"                  
#> [119] "Nepal"                    "New Zealand"             
#> [121] "Oman"                     "Pakistan"                
#> [123] "Panama"                   "Peru"                    
#> [125] "Philippines"              "Papua New Guinea"        
#> [127] "Poland"                   "Puerto Rico"             
#> [129] "North Korea"              "Portugal"                
#> [131] "Paraguay"                 "Palestine"               
#> [133] "Qatar"                    "Romania"                 
#> [135] "Russia"                   "Rwanda"                  
#> [137] "Saudi Arabia"             "Sudan"                   
#> [139] "Senegal"                  "Singapore"               
#> [141] "Solomon Islands"          "Sierra Leone"            
#> [143] "El Salvador"              "Serbia"                  
#> [145] "South Sudan"              "São Tomé & Príncipe"     
#> [147] "Suriname"                 "Slovakia"                
#> [149] "Slovenia"                 "Sweden"                  
#> [151] "Eswatini"                 "Syria"                   
#> [153] "Chad"                     "Togo"                    
#> [155] "Thailand"                 "Tajikistan"              
#> [157] "Turkmenistan"             "Timor-Leste"             
#> [159] "Tonga"                    "Trinidad & Tobago"       
#> [161] "Tunisia"                  "Turkey"                  
#> [163] "Tanzania"                 "Uganda"                  
#> [165] "Ukraine"                  "Uruguay"                 
#> [167] "US"                       "Uzbekistan"              
#> [169] "St. Vincent & Grenadines" "Venezuela"               
#> [171] "Vietnam"                  "Vanuatu"                 
#> [173] "Samoa"                    "Yemen"                   
#> [175] "South Africa"             "Zambia"                  
#> [177] "Zimbabwe"